[Pcl] Convert PCL target selector to dialog
