+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.csproj
@@ -7708,6 +7708,7 @@
     <Compile Include="MonoDevelop.Ide.Gui\DisplayBindingService.cs" />
     <Compile Include="MonoDevelop.Ide.Gui\BackgroundProgressMonitor.cs" />
     <Compile Include="MonoDevelop.Ide.Gui\StatusProgressMonitor.cs" />
+    <Compile Include="MonoDevelop.Ide.Projects.OptionPanels\PortableRuntimeSelectorDialog.cs" />
     <Compile Include="MonoDevelop.Ide.Tasks\TaskService.cs" />
     <Compile Include="MonoDevelop.Ide.Codons\ContextPadCodon.cs" />
     <Compile Include="MonoDevelop.Ide.Codons\WorkbenchContextCodon.cs" />
